refactor(auto api): renames and shifts
now the plugin is in one file, decluttering the vite.config file, and
gives it a better name
